+++ spec/search_flip/connection_spec.rb
@@ -95,6 +95,12 @@ RSpec.describe SearchFlip::Connection do
       expect(connection.get_indices.map { |index| index["index"] }.to_set).to eq(["comments", "products"].to_set)
       expect(connection.get_indices("com*").map { |index| index["index"] }).to eq(["comments"])
     end
+
+    it "accepts additional parameters" do
+      connection = SearchFlip::Connection.new
+
+      expect(connection.get_indices("comments", params: { h: "i,sth" })).to eq("i" => "comments", "sth" => "false")
+    end
   end
 
   describe "#create_index" do
@@ -154,6 +160,41 @@ RSpec.describe SearchFlip::Connection do
         connection.open_index("index_name")
 
         expect(connection.get_indices("index_name").first["status"]).to eq("open")
+      ensure
+        connection.delete_index("index_name") if connection.index_exists?("index_name")
+      end
+    end
+  end
+
+  describe "#freeze_index" do
+    it "freezes the specified index" do
+      begin
+        connection = SearchFlip::Connection.new
+
+        return if connection.version.to_f < 6.6
+
+        connection.create_index("index_name")
+        connection.freeze_index("index_name")
+
+        expect(connection.get_indices("index_name", params: { h: "sth" }).first["sth"]).to eq("true")
+      ensure
+        connection.delete_index("index_name") if connection.index_exists?("index_name")
+      end
+    end
+  end
+
+  describe "#unfreeze_index" do
+    it "unfreezes the specified index" do
+      begin
+        connection = SearchFlip::Connection.new
+
+        return if connection.version.to_f < 6.6
+
+        connection.create_index("index_name")
+        connection.freeze_index("index_name")
+        connection.unfreeze_index("index_name")
+
+        expect(connection.get_indices("index_name", params: { h: "sth" }).first["sth"]).to eq("false")
       ensure
         connection.delete_index("index_name") if connection.index_exists?("index_name")
       end