+++ spec/search_flip/index_spec.rb
@@ -95,10 +95,9 @@ RSpec.describe SearchFlip::Index do
 
   describe ".close_index" do
     it "delegates to connection" do
-      allow(TestIndex.connection).to receive(:close_index).and_call_original
+      allow(TestIndex.connection).to receive(:close_index)
 
       TestIndex.create_index
-      sleep(0.1) while TestIndex.connection.cluster_health["status"] == "red"
       TestIndex.close_index
 
       expect(TestIndex.connection).to have_received(:close_index).with("test")
@@ -107,15 +106,36 @@ RSpec.describe SearchFlip::Index do
 
   describe ".open_index" do
     it "delegates to connection" do
-      allow(TestIndex.connection).to receive(:open_index).and_call_original
+      allow(TestIndex.connection).to receive(:open_index)
 
       TestIndex.create_index
-      sleep(0.1) while TestIndex.connection.cluster_health["status"] == "red"
       TestIndex.close_index
-
       TestIndex.open_index
 
       expect(TestIndex.connection).to have_received(:open_index).with("test")
+    end
+  end
+
+  describe ".freeze_index" do
+    it "delegates to connection" do
+      allow(TestIndex.connection).to receive(:freeze_index)
+
+      TestIndex.create_index
+      TestIndex.freeze_index
+
+      expect(TestIndex.connection).to have_received(:freeze_index).with("test")
+    end
+  end
+
+  describe ".unfreeze_index" do
+    it "delegates to connection" do
+      allow(TestIndex.connection).to receive(:unfreeze_index)
+
+      TestIndex.create_index
+      TestIndex.freeze_index
+      TestIndex.unfreeze_index
+
+      expect(TestIndex.connection).to have_received(:unfreeze_index).with("test")
     end
   end